Renamed test helper class
