+++ service/src/test/resources/core-case-data-store-api/callback-request.json
@@ -524,7 +524,7 @@
           "id": "ca15c188-14d2-49a3-ac61-fdd7c6c7dc9b",
           "value": {
             "fullName": "Barney Rubble",
-            "positionInACase": "Cafcass Guardian",
+            "positionInACase": "Cafcass Solicitor",
             "address": {
               "County": "Galway",
               "Country": "Ireland",
